Emit set-course event to set course from visual steering
Had to make mounted life-cycle more robust also by moving
the v-if out from the component. This is quite annoying as
it sort of breaks logical units into two files but there's
no better way to avoid component initialization it seems
(at least created triggered some errors because the boat
is not yet in a valid position).