Block API fetching restart on subsequent calls
They would not harm but lets not play with the normal logic
here too much.

Wx update times change might be interesting in theory to catch
too but is not that frequently (never?) done. And even then the
latency in fetching the wx is just a cold period one so not a
big deal to let the normal logic to deal with it without any
forced extra refetching from here.