Refetch in-race messages less frequently
During testing of the new client, it was noticed that boat API
may update lmi before the in-race message is available through
the racemessages API, causing 1s retry timer to try many times
to no-avail. Increase the timer and make it random to avoid
potential synchronization problems between all clients.