Fix overly long popup content with Chrome
The old approach didn't create scrollbar with Chrome as needed.

The boat list creation popup is now sub-optimally sized with smaller
screens but it's better to avoid the UI issue than have it sized
perfectly. Maybe CSS can also make its lists shrink when needed but
currently it's not the case.