Add extra debug for the delay value (0.1) rejected problem
This problem was reported and it is unreproducable for me.
The code looks too obvious to fail.

Only thing I could think of is some kind of locale related problem
rejecting 0,1 but even that should return 0 rather than NaN so
better to try to figure it out the hard way.