Degrade to chatless mode if boat API fails w/ chatroom fetching
Detect persisting failures in fetching chat data that is more
risky by definition than normal boat api data. The latter is
mission critical for the simulator to client communications and
it may be useful to drop chat refreshing if it is causing
problems to allow boat API to function otherwise normally.