Sort polar curves based on distance to the current tws
With >30kn polar curves, they're rarely wanted to be the topmost
curves so drawing in order was somewhat suboptimal as they seemed
to hide some more useful bits of the polar here and there.

This change makes to polar curve ordering depend on the current
TWS. It will, however, impact "the usual look" of a polar because
the general appearance will typically be less red/orange.